[Dashboard Agent] extract dashboard schema types
Move dashboard schema definitions into a dedicated shared module and keep the attachment types file focused on attachment wrappers and re-exports. This further separates non-functional API cleanup from the harder behavior changes that will follow.

Made-with: Cursor
